pyproject.toml
[project]
name = "telekomexpport2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.30.0",
    "pandas>=1.5.0",
    "rich>=12.0.0",
    "pydantic>=1.10.0",
    "python-dotenv>=0.21.0",
    "matplotlib>=3.5.0",
    "tabulate>=0.9.0",
    "pyotp>=2.9.0",
]
